修改学生报名序号逻辑为redis中获取； 统一redis配置

+++ src/main/java/com/mx/ai/sports/course/service/ICourseStudentService.java
@@ -105,7 +105,7 @@ public interface ICourseStudentService extends IService<CourseStudent> {
      * @param courseId
      * @return
      */
-    Long findEntryStudentList2Redis(Long courseId, Long userId);
+    Integer findEntryStudentList2Redis(Long courseId, Long userId);
 
     /**
      * 删除报名课程的学生列表
